Abstraction: separating extensions from Window
Trouble:
I need to initiate them explicitly like

```cs
        public byte Main(object prm)
        {
            byte a = States.IsActuallyOk(((iWinBuff)this).MainBuff(prm));
            if(a != States.OK) return a;
            a |= States.IsActuallyOk(((iWinLoop)this).MainLoop(prm));
            if(a != States.OK) return a;
            a |= States.IsActuallyOk(((iWinPtr)this).MainPtr(prm));
            return 0;
        }
```

+++ test/App.axaml.cs
@@ -15,7 +15,7 @@ namespace test
         {
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
             {
-                int a;
+                byte a;
                 desktop.MainWindow = new MinCtrlWindow(out a);
             }
